--- a/osu.Game/Beatmaps/BeatmapInfoExtensions.cs
+++ b/osu.Game/Beatmaps/BeatmapInfoExtensions.cs
@@ -1,8 +1,8 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using System.Collections.Generic;
 using osu.Framework.Localisation;
+using osu.Game.Screens.Select;
 
 namespace osu.Game.Beatmaps
 {
@@ -29,20 +29,18 @@ namespace osu.Game.Beatmaps
             return new RomanisableString($"{metadata.GetPreferred(true)}".Trim(), $"{metadata.GetPreferred(false)}".Trim());
         }
 
-        public static List<string> GetSearchableTerms(this IBeatmapInfo beatmapInfo)
+        public static bool Match(this IBeatmapInfo beatmapInfo, params FilterCriteria.OptionalTextFilter[] filters)
         {
-            var termsList = new List<string>(BeatmapMetadataInfoExtensions.MAX_SEARCHABLE_TERM_COUNT + 1);
-
-            addIfNotNull(beatmapInfo.DifficultyName);
-
-            BeatmapMetadataInfoExtensions.CollectSearchableTerms(beatmapInfo.Metadata, termsList);
-            return termsList;
-
-            void addIfNotNull(string? s)
+            foreach (var filter in filters)
             {
-                if (!string.IsNullOrEmpty(s))
-                    termsList.Add(s);
+                if (filter.Matches(beatmapInfo.DifficultyName))
+                    return true;
+
+                if (BeatmapMetadataInfoExtensions.Match(beatmapInfo.Metadata, filters))
+                    return true;
             }
+
+            return false;
         }
 
         private static string getVersionString(IBeatmapInfo beatmapInfo) => string.IsNullOrEmpty(beatmapInfo.DifficultyName) ? string.Empty : $"[{beatmapInfo.DifficultyName}]";