Just a quick optimisation pass. Anything further will require more consideration, as the overhead is coming from the raw `IndexTo` operation, and to a lesser extent `CarouselBeatmap.checkMatch()` and other methods with nothing too simple to optimise out:

![CleanShot 2023-12-05 at 06 32 00](https://github.com/ppy/osu/assets/191335/4a0e2d1e-712b-42cb-9303-c842f9dcac1d)

Initial filter is the full load of song select (includes sort). Subsequent are filtering on non-matching strings (ie. "aaaaaaa").

`master`:

[runtime] 2023-12-05 05:23:02 [verbose]: Filter took 8,785.0 ms
[runtime] 2023-12-05 05:23:39 [verbose]: Filter took 9,147.0 ms
[runtime] 2023-12-05 05:23:50 [verbose]: Filter took 9,159.0 ms
[runtime] 2023-12-05 05:24:00 [verbose]: Filter took 9,254.0 ms
[runtime] 2023-12-05 05:24:16 [verbose]: Filter took 9,474.0 ms

after 7fda38d0b02496f4e255e496a73def950deca6a3:

[runtime] 2023-12-05 05:20:36 [verbose]: Filter took 8,739.0 ms
[runtime] 2023-12-05 05:21:40 [verbose]: Filter took 6,354.0 ms
[runtime] 2023-12-05 05:21:47 [verbose]: Filter took 6,471.0 ms
[runtime] 2023-12-05 05:21:53 [verbose]: Filter took 6,437.0 ms
[runtime] 2023-12-05 05:22:02 [verbose]: Filter took 6,753.0 ms

after 27e778ae09364981ce5e2768110efde51b4a3f7b:

[runtime] 2023-12-05 05:18:30 [verbose]: Filter took 8,990.0 ms
[runtime] 2023-12-05 05:19:15 [verbose]: Filter took 1,266.0 ms
[runtime] 2023-12-05 05:19:17 [verbose]: Filter took 1,213.0 ms
[runtime] 2023-12-05 05:19:19 [verbose]: Filter took 1,179.0 ms
[runtime] 2023-12-05 05:19:21 [verbose]: Filter took 1,166.0 ms

after 36eb338f53db580eafc8f2a576fcf8c4153a3173:

[runtime] 2023-12-05 05:54:19 [verbose]: Filter took 8,789.0 ms
[runtime] 2023-12-05 05:54:41 [verbose]: Filter took 1,232.0 ms
[runtime] 2023-12-05 05:54:43 [verbose]: Filter took 1,061.0 ms
[runtime] 2023-12-05 05:54:44 [verbose]: Filter took 1,039.0 ms
[runtime] 2023-12-05 05:54:46 [verbose]: Filter took 1,032.0 ms

Testing done using:

```diff
diff --git a/osu.Game/Screens/Select/BeatmapCarousel.cs b/osu.Game/Screens/Select/BeatmapCarousel.cs
index eb47a7201a..53bfef7d87 100644
--- a/osu.Game/Screens/Select/BeatmapCarousel.cs
+++ b/osu.Game/Screens/Select/BeatmapCarousel.cs
@@ -17,6 +17,7 @@
 using osu.Framework.Input.Bindings;
 using osu.Framework.Input.Events;
 using osu.Framework.Layout;
+using osu.Framework.Logging;
 using osu.Framework.Threading;
 using osu.Framework.Utils;
 using osu.Game.Beatmaps;
@@ -131,6 +132,9 @@ private void loadBeatmapSets(IEnumerable<BeatmapSetInfo> beatmapSets)
         {
             originalBeatmapSetsDetached = beatmapSets.Detach();
 
+            while (originalBeatmapSetsDetached.Count < 250000)
+                originalBeatmapSetsDetached.AddRange(originalBeatmapSetsDetached);
+
             if (selectedBeatmapSet != null && !originalBeatmapSetsDetached.Contains(selectedBeatmapSet.BeatmapSet))
                 selectedBeatmapSet = null;
 
@@ -738,6 +742,9 @@ void perform()
                     return;
                 }
 
+                Stopwatch sw = new Stopwatch();
+                sw.Start();
+
                 root.Filter(activeCriteria);
                 itemsCache.Invalidate();
 
@@ -745,6 +752,8 @@ void perform()
                     ScrollToSelected(true);
 
                 FilterApplied?.Invoke();
+
+                Logger.Log($"Filter took {sw.ElapsedMilliseconds:N1} ms");
             }
         }
 

```

Improves https://github.com/ppy/osu/issues/11571 but probably doesn't close yet (would want things to be async still).


---

Reduces time taken to search beatmaps at song select by 88%.